fix a lot

+++ controllers/post.controller.js
@@ -194,11 +194,21 @@ exports.updatePost = async (req, res) => {
 // delete post
 exports.deletePost = async (req, res) => {
 	const { id } = req.params;
+	const { userId } = req.body;
+	if (!userId) return res.status(404).json({ message: 'No user ID found', type: 'error' });
 
 	try {
 		// find and update
+		let user = await User.findById(userId).populate([{ path: 'saved', select: '_id' }]);
+
+		// loai post khoi user.saved
+		if (user.saved.some((f) => f._id.toString() === id)) {
+			user.saved = user.saved.filter((f) => f._id.toString() !== id);
+		}
+
+		// find and update
 		await Post.findByIdAndDelete(id);
-		return res.status(200).json({ message: `Deleted post`, type: 'success' });
+		return res.status(200).json({ message: `Deleted post`, type: 'success', user });
 	} catch (error) {
 		return res.status(500).json({ message: error.message, type: 'error' });
 	}
@@ -225,9 +235,13 @@ exports.save = async (req, res) => {
 
 			// lay lai user
 			const newUser = await User.findById(userId)
-				.populate([{ path: 'saved' }])
+				.populate([
+					{
+						path: 'saved',
+						populate: { path: 'author', select: '_id username profilePicture' },
+					},
+				])
 				.lean();
-
 			return res.status(200).json({ message: 'Saved this post', type: 'success', newUser });
 		} else {
 			return res
@@ -258,7 +272,12 @@ exports.unsave = async (req, res) => {
 
 			// lay lai user
 			const newUser = await User.findById(userId)
-				.populate([{ path: 'saved' }])
+				.populate([
+					{
+						path: 'saved',
+						populate: { path: 'author', select: '_id username profilePicture' },
+					},
+				])
 				.lean();
 
 			return res.status(200).json({ message: 'Unsave this post', type: 'success', newUser });
